+++ coderd/organizations_test.go
@@ -25,24 +25,6 @@ func TestOrganizationByUserAndName(t *testing.T) {
 		require.Equal(t, http.StatusNotFound, apiErr.StatusCode())
 	})
 
-	t.Run("NoMember", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		first := coderdtest.CreateFirstUser(t, client)
-		other, _ := coderdtest.CreateAnotherUser(t, client, first.OrganizationID)
-		ctx := testutil.Context(t, testutil.WaitLong)
-
-		org, err := client.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name:        "another",
-			DisplayName: "Another",
-		})
-		require.NoError(t, err)
-		_, err = other.OrganizationByUserAndName(ctx, codersdk.Me, org.Name)
-		var apiErr *codersdk.Error
-		require.ErrorAs(t, err, &apiErr)
-		require.Equal(t, http.StatusNotFound, apiErr.StatusCode())
-	})
-
 	t.Run("Valid", func(t *testing.T) {
 		t.Parallel()
 		client := coderdtest.New(t, nil)
@@ -56,70 +38,3 @@ func TestOrganizationByUserAndName(t *testing.T) {
 	})
 }
 
-func TestPostOrganizationsByUser(t *testing.T) {
-	t.Parallel()
-	t.Run("Conflict", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		user := coderdtest.CreateFirstUser(t, client)
-		ctx := testutil.Context(t, testutil.WaitLong)
-
-		org, err := client.Organization(ctx, user.OrganizationID)
-		require.NoError(t, err)
-		_, err = client.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name:        org.Name,
-			DisplayName: org.DisplayName,
-		})
-		var apiErr *codersdk.Error
-		require.ErrorAs(t, err, &apiErr)
-		require.Equal(t, http.StatusConflict, apiErr.StatusCode())
-	})
-
-	t.Run("InvalidName", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		_ = coderdtest.CreateFirstUser(t, client)
-		ctx := testutil.Context(t, testutil.WaitLong)
-
-		_, err := client.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name:        "A name which is definitely not url safe",
-			DisplayName: "New",
-		})
-		var apiErr *codersdk.Error
-		require.ErrorAs(t, err, &apiErr)
-		require.Equal(t, http.StatusBadRequest, apiErr.StatusCode())
-	})
-
-	t.Run("Create", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		_ = coderdtest.CreateFirstUser(t, client)
-		ctx := testutil.Context(t, testutil.WaitLong)
-
-		o, err := client.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name:        "new-org",
-			DisplayName: "New organization",
-			Description: "A new organization to love and cherish forever.",
-			Icon:        "/emojis/1f48f-1f3ff.png",
-		})
-		require.NoError(t, err)
-		require.Equal(t, "new-org", o.Name)
-		require.Equal(t, "New organization", o.DisplayName)
-		require.Equal(t, "A new organization to love and cherish forever.", o.Description)
-		require.Equal(t, "/emojis/1f48f-1f3ff.png", o.Icon)
-	})
-
-	t.Run("CreateWithoutExplicitDisplayName", func(t *testing.T) {
-		t.Parallel()
-		client := coderdtest.New(t, nil)
-		_ = coderdtest.CreateFirstUser(t, client)
-		ctx := testutil.Context(t, testutil.WaitLong)
-
-		o, err := client.CreateOrganization(ctx, codersdk.CreateOrganizationRequest{
-			Name: "new-org",
-		})
-		require.NoError(t, err)
-		require.Equal(t, "new-org", o.Name)
-		require.Equal(t, "new-org", o.DisplayName) // should match the given `Name`
-	})
-}
